Separate QueryResolver and QueryValidator into separate files. Remove the 'implicit' keyword for toUnresolved type converter to make them normal methods.
